Move getDownloadCommand
